Correccion de bug: se cambia int por double, redondeaba siempre a cero la prioridad

+++ projects/arco-iris/arco-iris-core/src/main/java/ar/uba/dc/arcoiris/selfhealing/priority/ScenarioRelativePriorityAssigner.java
@@ -4,6 +4,6 @@ import ar.uba.dc.arcoiris.selfhealing.SelfHealingScenario;
 
 public interface ScenarioRelativePriorityAssigner {
 
-	public abstract int relativePriority(SelfHealingScenario scenario);
+	public abstract double relativePriority(SelfHealingScenario scenario);
 
 }
